Generate new seed on scene start
When the scene was run it always took the current seed that was set in
the editor. Now it always used a different seed. This change was made to
make running the scene more interesting and because I always forgot to
change the seed and it felt like a chore. Maybe later it should be added
as a variable like "d_new_seed_on_start" or something like this.