Add checking for editor & fix heightmap sampling
The check for editor is added to avoid calling set_owner() if it's not
running in the editor.

I also accidentaly included a fix for a heightmap-sampling bug in the
last commit (a075eee7913cacbd5ebca74ce21fd0ebbfce0d11):

In scripts/TerrainGenerator.gd (lines: 439 & 440) I added the
terrain_unit_size because the heightmap-sampling did not work properly
when increasing the value of terrain_unit_size.