Fix scene not starting
The heightmap-sampling takes too long and crashes my Godot install after
some time. I think I have to optimize it somehow..

Because of this I also disabled generating the terrain after loading the
scene inside of the editor. Had a few crashes directly after opening the
project because of this..