Update TEST-STUFF position & add debug messages
The TEST-STUFF node was below the terrain mesh most of the time so the
camera didn't show anything..

I also removed the _id_editor condition in _ready() because it broke
the mesh generation when running the current scene because it still was
in run through the editor.

Furthermore I added a debug message to see which chunk # is currently
generating.